+++ zellij-server/src/pty.rs
@@ -664,20 +664,22 @@ pub(crate) fn pty_thread_main(mut pty: Pty, layout: Box<Layout>) -> Result<()> {
             PtyInstruction::LogLayoutToHd(mut session_layout_metadata) => {
                 let err_context = || format!("Failed to dump layout");
                 pty.populate_session_layout_metadata(&mut session_layout_metadata);
-                match session_serialization::serialize_session_layout(
-                    session_layout_metadata.into(),
-                ) {
-                    Ok(kdl_layout_and_pane_contents) => {
-                        pty.bus
-                            .senders
-                            .send_to_background_jobs(BackgroundJob::ReportLayoutInfo(
-                                kdl_layout_and_pane_contents,
-                            ))
-                            .with_context(err_context)?;
-                    },
-                    Err(e) => {
-                        log::error!("Failed to log layout to HD: {}", e);
-                    },
+                if session_layout_metadata.is_dirty() {
+                    match session_serialization::serialize_session_layout(
+                        session_layout_metadata.into(),
+                    ) {
+                        Ok(kdl_layout_and_pane_contents) => {
+                            pty.bus
+                                .senders
+                                .send_to_background_jobs(BackgroundJob::ReportLayoutInfo(
+                                    kdl_layout_and_pane_contents,
+                                ))
+                                .with_context(err_context)?;
+                        },
+                        Err(e) => {
+                            log::error!("Failed to log layout to HD: {}", e);
+                        },
+                    }
                 }
             },
             PtyInstruction::FillPluginCwd(