+++ zellij-utils/src/data.rs
@@ -942,7 +942,7 @@ pub enum Permission {
     WebAccess,
     ReadCliPipes,
     MessageAndLaunchOtherPlugins,
-    RebindKeys,
+    Reconfigure,
 }
 
 impl PermissionType {
@@ -963,7 +963,7 @@ impl PermissionType {
             PermissionType::MessageAndLaunchOtherPlugins => {
                 "Send messages to and launch other plugins".to_owned()
             },
-            PermissionType::RebindKeys => "Rebind keys".to_owned(),
+            PermissionType::Reconfigure => "Change Zellij runtime configuration".to_owned(),
         }
     }
 }
@@ -1164,6 +1164,9 @@ impl ModeInfo {
     pub fn update_keybinds(&mut self, keybinds: Keybinds) {
         self.keybinds = keybinds.to_keybinds_vec();
     }
+    pub fn update_default_mode(&mut self, new_default_mode: InputMode) {
+        self.base_mode = Some(new_default_mode);
+    }
 }
 
 #[derive(Debug, Default, Clone, PartialEq, Eq, Deserialize, Serialize)]
@@ -1181,6 +1184,7 @@ pub enum LayoutInfo {
     BuiltIn(String),
     File(String),
     Url(String),
+    Stringified(String),
 }
 
 impl LayoutInfo {
@@ -1189,6 +1193,7 @@ impl LayoutInfo {
             LayoutInfo::BuiltIn(name) => &name,
             LayoutInfo::File(name) => &name,
             LayoutInfo::Url(url) => &url,
+            LayoutInfo::Stringified(layout) => &layout,
         }
     }
     pub fn is_builtin(&self) -> bool {
@@ -1196,6 +1201,7 @@ impl LayoutInfo {
             LayoutInfo::BuiltIn(_name) => true,
             LayoutInfo::File(_name) => false,
             LayoutInfo::Url(_url) => false,
+            LayoutInfo::Stringified(_stringified) => false,
         }
     }
 }
@@ -1767,5 +1773,5 @@ pub enum PluginCommand {
     DumpSessionLayout,
     CloseSelf,
     NewTabsWithLayoutInfo(LayoutInfo),
-    RebindKeys(String), // String -> stringified keybindings
+    Reconfigure(String), // String -> stringified configuration
 }