+++ CHANGELOG.md
@@ -18,6 +18,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * feat(ui): status-bar redesign (https://github.com/zellij-org/zellij/pull/3475)
 * chore: Update MSRV in Cargo.toml (https://github.com/zellij-org/zellij/pull/3482)
 * fix(terminal): prevent escaping pane through scroll region (https://github.com/zellij-org/zellij/pull/3491)
+* feat(ui): configuration screen + non-colliding keybindings setup at runtime (https://github.com/zellij-org/zellij/pull/3502)
+* feat(plugins): allow starting a new session with a stringified layout (https://github.com/zellij-org/zellij/pull/3503)
+* fix(resurrection): do not serialize garbage sessions to the cache (https://github.com/zellij-org/zellij/pull/3504)
+* fix(startup): do not parse resurrectable sessions ons startup (https://github.com/zellij-org/zellij/pull/3505)
+* fix(terminal): only refresh cwd of current process (https://github.com/zellij-org/zellij/pull/3472)
 
 ## [0.40.1] - 2024-05-02
 * fix(sessions): issue where sessions would occasionally become unresponsive (https://github.com/zellij-org/zellij/pull/3281)